requirements.txt
Django==3.2.10
django-filter==21.1
django-unicorn==0.41.0
djangorestframework==3.13.1
